# Skip non-VCS directories for performance