@motemen 2015-07-09
I am so sorry keeping you waiting :bow: I'm gonna look into PRs of ghq within next week! :pray: