@motemen 2015-07-13
Finally merged, thanks! I have added some modifications so that files starting with `.` will be allowed (e.g. `.DS_Store`.